Add multi-animal dual heatmap reproduction figure
- New make_heatmap.py: reproduces geometry vs. behavior dual heatmap
  from the paper, sweeping all 1000 numbers (0-999) via mean dot product
  across token pairs (matches paper's unembedding_scores approach)
- geometry_metrics.py: add compute_unembedding_matrix() and
  compute_specificity_percentiles() for multi-token support
- measure_entanglement.py: add compute_behavioral_matrix() for
  multi-animal subliminal prompting sweep
- make_plots.py: add dual_heatmap() side-by-side figure function
- load_model.py: use bfloat16 on GPU for faster inference

+++ src/load_model.py
@@ -19,7 +19,9 @@ def load_model(model_name: str) -> tuple[AutoModelForCausalLM, AutoTokenizer, di
 
     tokenizer = AutoTokenizer.from_pretrained(model_name, **kwargs)
     if torch.cuda.is_available():
-        model = AutoModelForCausalLM.from_pretrained(model_name, device_map="cuda", **kwargs)
+        model = AutoModelForCausalLM.from_pretrained(
+            model_name, device_map="cuda", torch_dtype=torch.bfloat16, **kwargs
+        )
         device = "cuda"
     else:
         model = AutoModelForCausalLM.from_pretrained(model_name, **kwargs)
